pyproject.toml
[tool.poetry]
name = "src"
version = "0.1.0"
description = ""
authors = ["AlexAdedeji <alexadedeji15@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic-settings = "^2.7.1"
sqlalchemy = "^2.0.37"
loguru = "^0.7.3"
tenacity = "^9.0.0"
psycopg2-binary = "^2.9.10"
psycopg2 = "^2.9.10"
motor = "^3.6.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pyjwt = "^2.10.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
black = "^24.10.0"
flake8 = "^7.1.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
